refactor: 히어릿 조회 정렬 기준에 대한 예외를 enum 안에서 검증하도록 변경 #633
