+++ backend/app/src/main/java/com/onair/hearit/app/hearit/dto/HearitSortRequest.java
@@ -1,5 +1,7 @@
 package com.onair.hearit.app.hearit.dto;
 
+import com.onair.hearit.app.hearit.dto.param.HearitSortField;
+import com.onair.hearit.app.hearit.dto.param.SortDirection;
 import org.springframework.data.domain.Sort;
 
 public record HearitSortRequest(HearitSortField field, SortDirection direction) {
