--- backend/app/src/main/java/com/onair/hearit/app/hearit/dto/SortDirection.java
@@ -1,18 +1,0 @@
-package com.onair.hearit.app.hearit.dto;
-
-import org.springframework.data.domain.Sort.Direction;
-
-public enum SortDirection {
-    ASC(Direction.ASC),
-    DESC(Direction.DESC);
-
-    private final Direction direction;
-
-    SortDirection(Direction direction) {
-        this.direction = direction;
-    }
-
-    public Direction getDirection() {
-        return direction;
-    }
-}
