+++ backend/app/src/test/java/com/onair/hearit/app/hearit/application/HearitServiceTest.java
@@ -10,10 +10,10 @@ import com.onair.hearit.app.exception.custom.NotFoundException;
 import com.onair.hearit.app.fixture.DbHelper;
 import com.onair.hearit.app.hearit.dto.HearitDetailResponse;
 import com.onair.hearit.app.hearit.dto.FilteredHearitResponse;
-import com.onair.hearit.app.hearit.dto.HearitSortField;
+import com.onair.hearit.app.hearit.dto.param.HearitSortField;
 import com.onair.hearit.app.hearit.dto.HearitSortRequest;
 import com.onair.hearit.app.hearit.dto.HearitsWithRecommendCategoryResponse;
-import com.onair.hearit.app.hearit.dto.SortDirection;
+import com.onair.hearit.app.hearit.dto.param.SortDirection;
 import com.onair.hearit.core.domain.Bookmark;
 import com.onair.hearit.core.domain.Category;
 import com.onair.hearit.core.domain.Hearit;
